Created an extension on Double to enable the formatting of currency and percentages.  Also added logic to set coin.priceChangePercentage24H in red or green depending on its increase or decrease

+++ CryptoCurrencySwiftUI/Models/Coin.swift
@@ -22,7 +22,7 @@ struct Coin: Codable, Identifiable {
     let currentPrice: Double
     let image: String?
     let marketCap: Double?
-    let marketCapRank: Double?
+    let marketCapRank: Int?
     let fullyDilutedValuation: Double?
     let totalVolume: Double?
     let high24H: Double?
